df: move Filesystem to filesystem.rs
Move the `Filesystem` struct into a new module `filesystem.rs`.
